zPAM3.1 BETA 6
New stuff:
- Full Team Deathmatch and Deathmatch gametype modes
- Fullscreen image of map with warning is showed when important cvars 
(like com_maxfps) are not in allowed range. (this is kind of replace of 
cvar checks when punkbuster is off)
- In readyup, sound "Time to go get movin" is played to last not-ready 
player (only if it is 3v3 or more).
- Prone peak fix (off by default) - when you go from prone to crouch 
position, you are allowed to prone again after a small delay. This 
should avoid these situations: you are proned, leaned left/right and you 
peak behind the wall by going to crouch and back to prone as fast as 
possible. For enemy its difficult to shoot, because character animation 
is bugged and moving too fast.

Changed stuff:
- Rebalanced shotgun is off by default in cg modes, in other modes is 
turned on.
- PB cvars and md5tools inside pbsv.cfg was cleaned a bit.